Move JEI and EMI interop into Common (#1206)
Implements #945, resolves #524
