+++ Common/src/main/java/at/petrak/hexcasting/interop/jei/PatternDrawable.java
@@ -1,4 +1,4 @@
-package at.petrak.hexcasting.fabric.interop.emi;
+package at.petrak.hexcasting.interop.jei;
 
 import at.petrak.hexcasting.api.casting.math.HexPattern;
 import at.petrak.hexcasting.api.mod.HexTags;
@@ -10,27 +10,23 @@ import at.petrak.hexcasting.client.render.PatternSettings.PositionSettings;
 import at.petrak.hexcasting.client.render.PatternSettings.StrokeSettings;
 import at.petrak.hexcasting.client.render.PatternSettings.ZappySettings;
 import at.petrak.hexcasting.xplat.IXplatAbstractions;
-import dev.emi.emi.api.render.EmiRenderable;
+import mezz.jei.api.gui.drawable.IDrawable;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.resources.ResourceLocation;
 
-public class PatternRendererEMI implements EmiRenderable {
-
+public class PatternDrawable implements IDrawable {
     private final int width;
     private final int height;
 
-    private int xOffset = 0;
-    private int yOffset = 0;
-
     private boolean strokeOrder;
-
     private final HexPattern pat;
+
     private PatternSettings patSets;
 
-    public PatternRendererEMI(ResourceLocation pattern, int w, int h) {
+    public PatternDrawable(ResourceLocation pattern, int w, int h) {
         var regi = IXplatAbstractions.INSTANCE.getActionRegistry();
         var entry = regi.get(pattern);
-        this.strokeOrder = HexUtils.isOfTag(regi, pattern, HexTags.Actions.PER_WORLD_PATTERN);
+        this.strokeOrder = !HexUtils.isOfTag(regi, pattern, HexTags.Actions.PER_WORLD_PATTERN);
         this.pat = entry.prototype();
         this.width = w;
         this.height = h;
@@ -41,29 +37,33 @@ public class PatternRendererEMI implements EmiRenderable {
                 ZappySettings.READABLE);
     }
 
-    public PatternRendererEMI shift(int x, int y) {
-        xOffset += x;
-        yOffset += y;
-        return this;
+    @Override
+    public int getWidth() {
+        return width;
     }
 
-    public PatternRendererEMI strokeOrder(boolean order) {
+    @Override
+    public int getHeight() {
+        return height;
+    }
+
+    public PatternDrawable strokeOrder(boolean order) {
         if(order != strokeOrder){
             patSets = new PatternSettings("pattern_drawable_" + width + "_" + height + (order ? "" : "nostroke"),
                     patSets.posSets,
                     patSets.strokeSets,
                     order ? ZappySettings.READABLE : ZappySettings.STATIC
-            );
+                    );
         }
         strokeOrder = order;
         return this;
     }
 
     @Override
-    public void render(GuiGraphics graphics, int x, int y, float delta) {
+    public void draw(GuiGraphics graphics, int x, int y) {
         var ps = graphics.pose();
         ps.pushPose();
-        ps.translate(xOffset + x, yOffset + y + 1, 0);
+        ps.translate(x, y + 1, 0);
         PatternRenderer.renderPattern(pat, graphics.pose(), patSets,
                 new PatternColors(0xc8_0c0a0c, 0xff_333030).withDotColors(0x80_666363, 0),
                 0, 10
