+++ Common/src/main/resources/assets/hexcasting/textures/gui/brainsweep_recipe_icon.png
