+++ Common/src/main/resources/assets/hexcasting/textures/gui/edify_recipe.png
