+++ Common/src/main/resources/assets/hexcasting/textures/gui/edify_recipe_icon.png
