+++ Common/src/main/resources/assets/hexcasting/textures/gui/phial_recipe.png
