+++ Fabric/build.gradle
@@ -58,16 +58,18 @@ loom {
 repositories {
     mavenCentral()
 
-    // paucal and patchi
-    maven { url = 'https://maven.blamejared.com' }
+    // paucal, patchi, inline, JEI
+    maven { url 'https://maven.blamejared.com' }
     // modmenu and clothconfig
     maven { url "https://maven.shedaniel.me/" }
     maven { url 'https://maven.ladysnake.org/releases' }
     // Entity reach
     maven { url "https://maven.jamieswhiteshirt.com/libs-release/" }
     maven { url "https://mvn.devos.one/snapshots/" }
-    // temporarily changed from https://maven.terraformersmc.com/releases/ due to intermittent errors
-    maven { url = "https://maven.gnomecraft.net/releases" }
+    // JEI fallback mirror
+    maven { url "https://modmaven.dev" }
+    // EMI (temporarily changed from https://maven.terraformersmc.com/releases/ due to intermittent errors)
+    maven { url "https://maven.gnomecraft.net/releases" }
     exclusiveContent {
         forRepository {
             maven { url = "https://api.modrinth.com/maven" }
@@ -125,6 +127,9 @@ dependencies {
 
     modImplementation "dev.emi:trinkets:$trinketsVersion"
     modImplementation "dev.emi:emi-fabric:${emiVersion}"
+
+    modCompileOnlyApi("mezz.jei:jei-${minecraftVersion}-fabric-api:${jeiVersion}")
+    modRuntimeOnly("mezz.jei:jei-${minecraftVersion}-fabric:${jeiVersion}")
 
 //    modImplementation "maven.modrinth:gravity-api:$gravityApiVersion"
     modApi("com.github.Virtuoel:Pehkui:${pehkuiVersion}", {
