+++ Forge/build.gradle
@@ -49,16 +49,12 @@ repositories {
         dir 'libs'
     }
 
-    maven {
-        // location of the maven that hosts JEI files
-        name = "Progwml6 maven"
-        url = "https://dvs1.progwml6.com/files/maven/"
-    }
-    maven {
-        // location of a maven mirror for JEI files, as a fallback
-        name = "ModMaven"
-        url = "https://modmaven.dev"
-    }
+    // paucal, patchi, inline, JEI
+    maven { url = 'https://maven.blamejared.com' }
+    // JEI fallback mirror
+    maven { url = "https://modmaven.dev" }
+    // EMI (temporarily changed from https://maven.terraformersmc.com/releases/ due to intermittent errors)
+    maven { url = "https://maven.gnomecraft.net/releases" }
     // caelus elytra
     maven { url = "https://maven.theillusivec4.top" }
     // pehkui
@@ -102,6 +98,7 @@ dependencies {
     compileOnly fg.deobf("mezz.jei:jei-$minecraftVersion-common-api:$jeiVersion")
     compileOnly fg.deobf("mezz.jei:jei-$minecraftVersion-forge-api:$jeiVersion")
     runtimeOnly fg.deobf("mezz.jei:jei-$minecraftVersion-forge:$jeiVersion")
+    compileOnly fg.deobf("dev.emi:emi-forge:$emiVersion+$minecraftVersion")
 
     compileOnly fg.deobf("top.theillusivec4.curios:curios-forge:$curiosVersion+$minecraftVersion:api")
     runtimeOnly fg.deobf("top.theillusivec4.curios:curios-forge:$curiosVersion+$minecraftVersion")
