+++ gradle.properties
@@ -17,6 +17,7 @@ paucalVersion=0.6.0-pre-118
 patchouliVersion=83
 
 jeiVersion=15.0.0.12
+emiVersion=1.0.4
 pehkuiVersion=3.8.2
 
 inlineVersion=1.2.2
